Basic Functionality
The board automatically cycles through listed players, accepts the next play on an empty square, and disables squares that have been played.

Option to manually adjust state on the Python side is incomplete. Manual changes are committed if the next interaction is with the component, but last manual change will revert if the next action is outside of the component.